legal_spend_main/requirements.txt
google-adk
pandas
pyarrow
pyodbc
openpyxl
streamlit
python-dotenv
litellm
ollama
numpy


